+++ nucypher/network/nodes.py
@@ -15,21 +15,529 @@ You should have received a copy of the GNU General Public License
 along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
 import os
+import random
+from collections import defaultdict
+from collections import deque
+from contextlib import suppress
+from logging import Logger
+from tempfile import TemporaryDirectory
+from typing import Set, Tuple
 
 import OpenSSL
 import maya
+import requests
+import time
 from constant_sorrow import constants
 from cryptography.hazmat.primitives.serialization import Encoding
 from cryptography.x509 import Certificate, NameOID
 from eth_keys.datatypes import Signature as EthSignature
+from requests.exceptions import SSLError
+from twisted.internet import reactor, defer
+from twisted.internet import task
+from twisted.internet.threads import deferToThread
 from twisted.logger import Logger
 
 from nucypher.config.constants import SeednodeMetadata
 from nucypher.config.keyring import _write_tls_certificate
+from nucypher.config.storages import InMemoryNodeStorage
+from nucypher.crypto.api import keccak_digest
 from nucypher.crypto.powers import BlockchainPower, SigningPower, EncryptingPower, NoSigningPower
+from nucypher.crypto.signing import signature_splitter
 from nucypher.network.middleware import RestMiddleware
+from nucypher.network.nicknames import nickname_from_seed
 from nucypher.network.protocols import SuspiciousActivity
 from nucypher.network.server import TLSHostingPower
+
+
+class Learner:
+    """
+    Any participant in the "learning loop" - a class inheriting from
+    this one has the ability, synchronously or asynchronously,
+    to learn about nodes in the network, verify some essential
+    details about them, and store information about them for later use.
+    """
+
+    _SHORT_LEARNING_DELAY = 5
+    _LONG_LEARNING_DELAY = 90
+    LEARNING_TIMEOUT = 10
+    _ROUNDS_WITHOUT_NODES_AFTER_WHICH_TO_SLOW_DOWN = 10
+
+    # For Keeps
+    __DEFAULT_NODE_STORAGE = InMemoryNodeStorage
+    __DEFAULT_MIDDLEWARE_CLASS = RestMiddleware
+
+    class NotEnoughTeachers(RuntimeError):
+        pass
+
+    class UnresponsiveTeacher(ConnectionError):
+        pass
+
+    def __init__(self,
+                 common_name: str,
+                 network_middleware: RestMiddleware = __DEFAULT_MIDDLEWARE_CLASS(),
+                 start_learning_now: bool = False,
+                 learn_on_same_thread: bool = False,
+                 known_nodes: tuple = None,
+                 seed_nodes: Tuple[tuple] = None,
+                 known_certificates_dir: str = None,
+                 node_storage=None,
+                 save_metadata: bool = False,
+                 abort_on_learning_error: bool = False
+                 ) -> None:
+
+        self.log = Logger("characters")  # type: Logger
+
+        self.__common_name = common_name
+        self.network_middleware = network_middleware
+        self.save_metadata = save_metadata
+        self.start_learning_now = start_learning_now
+        self.learn_on_same_thread = learn_on_same_thread
+
+        self._abort_on_learning_error = abort_on_learning_error
+        self._learning_listeners = defaultdict(list)
+        self._node_ids_to_learn_about_immediately = set()
+
+        self.known_certificates_dir = known_certificates_dir or TemporaryDirectory("nucypher-tmp-certs-").name
+        self.__known_nodes = FleetState()
+
+        # Read
+        if node_storage is None:
+            node_storage = self.__DEFAULT_NODE_STORAGE(federated_only=self.federated_only,
+                                                       # TODO: remove federated_only
+                                                       character_class=self.__class__)
+
+        self.node_storage = node_storage
+        if save_metadata and node_storage is constants.NO_STORAGE_AVAILIBLE:
+            raise ValueError("Cannot save nodes without a configured node storage")
+
+        known_nodes = known_nodes or tuple()
+        self.unresponsive_startup_nodes = list()  # TODO: Attempt to use these again later
+        for node in known_nodes:
+            try:
+                self.remember_node(node)
+            except self.UnresponsiveTeacher:
+                self.unresponsive_startup_nodes.append(node)
+
+        self.teacher_nodes = deque()
+        self._current_teacher_node = None  # type: Teacher
+        self._learning_task = task.LoopingCall(self.keep_learning_about_nodes)
+        self._learning_round = 0  # type: int
+        self._rounds_without_new_nodes = 0  # type: int
+        self._seed_nodes = seed_nodes or []
+        self.unresponsive_seed_nodes = set()
+
+        if self.start_learning_now:
+            self.start_learning_loop(now=self.learn_on_same_thread)
+
+    @property
+    def known_nodes(self):
+        return self.__known_nodes
+
+    def load_seednodes(self,
+                       read_storages: bool = True,
+                       retry_attempts: int = 3,
+                       retry_rate: int = 2,
+                       timeout=3):
+        """
+        Engage known nodes from storages and pre-fetch hardcoded seednode certificates for node learning.
+        """
+
+        def __attempt_seednode_learning(seednode_metadata, current_attempt=1):
+            from nucypher.characters.lawful import Ursula
+            self.log.debug(
+                "Seeding from: {}|{}:{}".format(seednode_metadata.checksum_address,
+                                                seednode_metadata.rest_host,
+                                                seednode_metadata.rest_port))
+
+            seed_node = Ursula.from_seednode_metadata(seednode_metadata=seednode_metadata,
+                                                      network_middleware=self.network_middleware,
+                                                      certificates_directory=self.known_certificates_dir,
+                                                      timeout=timeout,
+                                                      federated_only=self.federated_only)  # TODO: 466
+            if seed_node is False:
+                self.unresponsive_seed_nodes.add(seednode_metadata)
+            else:
+                self.unresponsive_seed_nodes.discard(seednode_metadata)
+                self.remember_node(seed_node)
+
+        for seednode_metadata in self._seed_nodes:
+            __attempt_seednode_learning(seednode_metadata=seednode_metadata)
+
+        if not self.unresponsive_seed_nodes:
+            self.log.info("Finished learning about all seednodes.")
+
+        if read_storages is True:
+            self.read_nodes_from_storage()
+
+        if not self.known_nodes:
+            self.log.warn("No seednodes were available after {} attempts".format(retry_attempts))
+            # TODO: Need some actual logic here for situation with no seed nodes (ie, maybe try again much later)
+
+    def read_nodes_from_storage(self) -> set:
+        stored_nodes = self.node_storage.all(federated_only=self.federated_only)  # TODO: 466
+        for node in stored_nodes:
+            self.remember_node(node)
+
+    def remember_node(self, node, force_verification_check=False):
+
+        if node == self:  # No need to remember self.
+            return
+
+        # First, determine if this is an outdated representation of an already known node.
+        with suppress(KeyError):
+            already_known_node = self.known_nodes[node.checksum_public_address]
+            if not node.timestamp > already_known_node.timestamp:
+                self.log.debug("Skipping already known node {}".format(already_known_node))
+                # This node is already known.  We can safely return.
+                return
+
+        node.save_certificate_to_disk(directory=self.known_certificates_dir, force=True)  # TODO: Verify before force?
+        certificate_filepath = node.get_certificate_filepath(certificates_dir=self.known_certificates_dir)
+        try:
+            node.verify_node(force=force_verification_check,
+                             network_middleware=self.network_middleware,
+                             accept_federated_only=self.federated_only,  # TODO: 466
+                             certificate_filepath=certificate_filepath)
+        except SSLError:
+            raise  # TODO
+        except (requests.exceptions.ConnectionError, requests.exceptions.ReadTimeout):
+            self.log.info("No Response from known node {}|{}".format(node.rest_interface, node.checksum_public_address))
+            raise self.UnresponsiveTeacher
+
+        listeners = self._learning_listeners.pop(node.checksum_public_address, tuple())
+        address = node.checksum_public_address
+
+        self.__known_nodes[address] = node
+
+        if self.save_metadata:
+            self.write_node_metadata(node=node)
+
+        self.log.info("Remembering {}, popping {} listeners.".format(node.checksum_public_address, len(listeners)))
+        for listener in listeners:
+            listener.add(address)
+        self._node_ids_to_learn_about_immediately.discard(address)
+
+        nodes_to_consider = list(self.known_nodes.values()) + [self]
+        sorted_nodes = sorted(nodes_to_consider , key=lambda n: n.timestamp, reverse=True)
+        self.known_nodes.checksum = keccak_digest(b"".join(bytes(n) for n in sorted_nodes)).hex()
+
+        # TODO: Probably not mutate all these foreign attrs - ideally maybe move this whole method up to FleetState.
+        latest_node = sorted_nodes[0]
+        self.known_nodes.most_recent_node_change = latest_node.timestamp
+        self.known_nodes.updated = maya.now()
+
+    def start_learning_loop(self, now=False):
+        if self._learning_task.running:
+            return False
+        elif now:
+            self.load_seednodes()
+            self._learning_task()  # Unhandled error might happen here.  TODO: Call this in a safer place.
+            self.learning_deferred = self._learning_task.start(interval=self._SHORT_LEARNING_DELAY)
+            self.learning_deferred.addErrback(self.handle_learning_errors)
+            return self.learning_deferred
+        else:
+            seeder_deferred = deferToThread(self.load_seednodes)
+            learner_deferred = self._learning_task.start(interval=self._SHORT_LEARNING_DELAY, now=now)
+            seeder_deferred.addErrback(self.handle_learning_errors)
+            learner_deferred.addErrback(self.handle_learning_errors)
+            self.learning_deferred = defer.DeferredList([seeder_deferred, learner_deferred])
+            return self.learning_deferred
+
+    def stop_learning_loop(self):
+        """
+        Only for tests at this point.  Maybe some day for graceful shutdowns.
+        """
+
+    def handle_learning_errors(self, *args, **kwargs):
+        failure = args[0]
+        if self._abort_on_learning_error:
+            self.log.critical("Unhandled error during node learning.  Attempting graceful crash.")
+            reactor.callFromThread(self._crash_gracefully, failure=failure)
+        else:
+            self.log.warn("Unhandled error during node learning: {}".format(failure.getTraceback()))
+            if not self._learning_task.running:
+                self.start_learning_loop()  # TODO: Consider a single entry point for this with more elegant pause and unpause.
+
+    def _crash_gracefully(self, failure=None):
+        """
+        A facility for crashing more gracefully in the event that an exception
+        is unhandled in a different thread, especially inside a loop like the learning loop.
+        """
+        self._crashed = failure
+        failure.raiseException()
+        self.log.critical("{} crashed with {}".format(self.__common_name, failure))
+
+    def shuffled_known_nodes(self):
+        nodes_we_know_about = list(self.__known_nodes.values())
+        random.shuffle(nodes_we_know_about)
+        self.log.info("Shuffled {} known nodes".format(len(nodes_we_know_about)))
+        return nodes_we_know_about
+
+    def select_teacher_nodes(self):
+        nodes_we_know_about = self.shuffled_known_nodes()
+
+        if not nodes_we_know_about:
+            raise self.NotEnoughTeachers("Need some nodes to start learning from.")
+
+        self.teacher_nodes.extend(nodes_we_know_about)
+
+    def cycle_teacher_node(self):
+        # To ensure that all the best teachers are availalble, first let's make sure
+        # that we have connected to all the seed nodes.
+        if self.unresponsive_seed_nodes:
+            self.log.info("Still have unresponsive seed nodes; trying again to connect.")
+            self.load_seednodes()  # Ideally, this is async and singular.
+
+        if not self.teacher_nodes:
+            self.select_teacher_nodes()
+        try:
+            self._current_teacher_node = self.teacher_nodes.pop()
+        except IndexError:
+            error = "Not enough nodes to select a good teacher, Check your network connection then node configuration"
+            raise self.NotEnoughTeachers(error)
+        self.log.info("Cycled teachers; New teacher is {}".format(self._current_teacher_node.checksum_public_address))
+
+    def current_teacher_node(self, cycle=False):
+        if cycle:
+            self.cycle_teacher_node()
+
+        if not self._current_teacher_node:
+            self.cycle_teacher_node()
+
+        teacher = self._current_teacher_node
+
+        return teacher
+
+    def learn_about_nodes_now(self, force=False):
+        if self._learning_task.running:
+            self._learning_task.reset()
+            self._learning_task()
+        elif not force:
+            self.log.warn(
+                "Learning loop isn't started; can't learn about nodes now.  You can override this with force=True.")
+        elif force:
+            self.log.info("Learning loop wasn't started; forcing start now.")
+            self._learning_task.start(self._SHORT_LEARNING_DELAY, now=True)
+
+    def keep_learning_about_nodes(self):
+        """
+        Continually learn about new nodes.
+        """
+        self.learn_from_teacher_node(eager=False)  # TODO: Allow the user to set eagerness?
+
+    def learn_about_specific_nodes(self, canonical_addresses: Set):
+        self._node_ids_to_learn_about_immediately.update(canonical_addresses)  # hmmmm
+        self.learn_about_nodes_now()
+
+    # TODO: Dehydrate these next two methods.
+
+    def block_until_number_of_known_nodes_is(self,
+                                             number_of_nodes_to_know: int,
+                                             timeout: int = 10,
+                                             learn_on_this_thread: bool = False):
+        start = maya.now()
+        starting_round = self._learning_round
+
+        while True:
+            rounds_undertaken = self._learning_round - starting_round
+            if len(self.__known_nodes) >= number_of_nodes_to_know:
+                if rounds_undertaken:
+                    self.log.info("Learned about enough nodes after {} rounds.".format(rounds_undertaken))
+                return True
+
+            if not self._learning_task.running:
+                self.log.warn("Blocking to learn about nodes, but learning loop isn't running.")
+            if learn_on_this_thread:
+                try:
+                    self.learn_from_teacher_node(eager=True)
+                except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectTimeout):
+                    # TODO: Even this "same thread" logic can be done off the main thread.
+                    self.log.warn("Teacher was unreachable.  No good way to handle this on the main thread.")
+
+            # The rest of the fucking owl
+            if (maya.now() - start).seconds > timeout:
+                if not self._learning_task.running:
+                    raise self.NotEnoughTeachers("Learning loop is not running.  Start it with start_learning().")
+                else:
+                    raise self.NotEnoughTeachers("After {} seconds and {} rounds, didn't find {} nodes".format(
+                        timeout, rounds_undertaken, number_of_nodes_to_know))
+            else:
+                time.sleep(.1)
+
+    def block_until_specific_nodes_are_known(self,
+                                             canonical_addresses: Set,
+                                             timeout=LEARNING_TIMEOUT,
+                                             allow_missing=0,
+                                             learn_on_this_thread=False):
+        start = maya.now()
+        starting_round = self._learning_round
+
+        while True:
+            if self._crashed:
+                return self._crashed
+            rounds_undertaken = self._learning_round - starting_round
+            if canonical_addresses.issubset(self.__known_nodes):
+                if rounds_undertaken:
+                    self.log.info("Learned about all nodes after {} rounds.".format(rounds_undertaken))
+                return True
+
+            if not self._learning_task.running:
+                self.log.warn("Blocking to learn about nodes, but learning loop isn't running.")
+            if learn_on_this_thread:
+                self.learn_from_teacher_node(eager=True)
+
+            if (maya.now() - start).seconds > timeout:
+
+                still_unknown = canonical_addresses.difference(self.__known_nodes)
+
+                if len(still_unknown) <= allow_missing:
+                    return False
+                elif not self._learning_task.running:
+                    raise self.NotEnoughTeachers("The learning loop is not running.  Start it with start_learning().")
+                else:
+                    raise self.NotEnoughTeachers(
+                        "After {} seconds and {} rounds, didn't find these {} nodes: {}".format(
+                            timeout, rounds_undertaken, len(still_unknown), still_unknown))
+
+            else:
+                time.sleep(.1)
+
+    def _adjust_learning(self, node_list):
+        """
+        Takes a list of new nodes, adjusts learning accordingly.
+
+        Currently, simply slows down learning loop when no new nodes have been discovered in a while.
+        TODO: Do other important things - scrub, bucket, etc.
+        """
+        if node_list:
+            self._rounds_without_new_nodes = 0
+            self._learning_task.interval = self._SHORT_LEARNING_DELAY
+        else:
+            self._rounds_without_new_nodes += 1
+            if self._rounds_without_new_nodes > self._ROUNDS_WITHOUT_NODES_AFTER_WHICH_TO_SLOW_DOWN:
+                self.log.info("After {} rounds with no new nodes, it's time to slow down to {} seconds.".format(
+                    self._ROUNDS_WITHOUT_NODES_AFTER_WHICH_TO_SLOW_DOWN,
+                    self._LONG_LEARNING_DELAY))
+                self._learning_task.interval = self._LONG_LEARNING_DELAY
+
+    def _push_certain_newly_discovered_nodes_here(self, queue_to_push, node_addresses):
+        """
+        If any node_addresses are discovered, push them to queue_to_push.
+        """
+        for node_address in node_addresses:
+            self.log.info("Adding listener for {}".format(node_address))
+            self._learning_listeners[node_address].append(queue_to_push)
+
+    def network_bootstrap(self, node_list: list) -> None:
+        for node_addr, port in node_list:
+            new_nodes = self.learn_about_nodes_now(node_addr, port)
+            self.__known_nodes.update(new_nodes)
+
+    def get_nodes_by_ids(self, node_ids):
+        for node_id in node_ids:
+            try:
+                # Scenario 1: We already know about this node.
+                return self.__known_nodes[node_id]
+            except KeyError:
+                raise NotImplementedError
+        # Scenario 2: We don't know about this node, but a nearby node does.
+        # TODO: Build a concurrent pool of lookups here.
+
+        # Scenario 3: We don't know about this node, and neither does our friend.
+
+    def write_node_metadata(self, node, serializer=bytes) -> str:
+        return self.node_storage.save(node=node)
+
+    def learn_from_teacher_node(self, eager=True):
+        """
+        Sends a request to node_url to find out about known nodes.
+        """
+        self._learning_round += 1
+
+        try:
+            current_teacher = self.current_teacher_node()
+        except self.NotEnoughTeachers as e:
+            self.log.warn("Can't learn right now: {}".format(e.args[0]))
+            return
+
+        rest_url = current_teacher.rest_interface  # TODO: Name this..?
+
+        # TODO: Do we really want to try to learn about all these nodes instantly?
+        # Hearing this traffic might give insight to an attacker.
+        if VerifiableNode in self.__class__.__bases__:
+            announce_nodes = [self]
+        else:
+            announce_nodes = None
+
+        unresponsive_nodes = set()
+        try:
+
+            # TODO: Streamline path generation
+            certificate_filepath = current_teacher.get_certificate_filepath(
+                certificates_dir=self.known_certificates_dir)
+            response = self.network_middleware.get_nodes_via_rest(url=rest_url,
+                                                                  nodes_i_need=self._node_ids_to_learn_about_immediately,
+                                                                  announce_nodes=announce_nodes,
+                                                                  certificate_filepath=certificate_filepath)
+        except requests.exceptions.ConnectionError as e:
+            unresponsive_nodes.add(current_teacher)
+            teacher_rest_info = current_teacher.rest_information()[0]
+
+            # TODO: This error isn't necessarily "no repsonse" - let's maybe pass on the text of the exception here.
+            self.log.info("No Response from teacher: {}:{}.".format(teacher_rest_info.host, teacher_rest_info.port))
+            self.cycle_teacher_node()
+            return
+
+        if response.status_code != 200:
+            raise RuntimeError("Bad response from teacher: {} - {}".format(response, response.content))
+
+        signature, nodes = signature_splitter(response.content, return_remainder=True)
+
+        # TODO: This doesn't make sense - a decentralized node can still learn about a federated-only node.
+        from nucypher.characters.lawful import Ursula
+        node_list = Ursula.batch_from_bytes(nodes, federated_only=self.federated_only)  # TODO: 466
+
+        new_nodes = []
+        for node in node_list:
+
+            if node.checksum_public_address in self.known_nodes or node.checksum_public_address == self.__common_name:
+                continue  # TODO: 168 Check version and update if required.
+
+            try:
+                if eager:
+                    certificate_filepath = current_teacher.get_certificate_filepath(
+                        certificates_dir=self.known_certificates_dir)
+                    node.verify_node(self.network_middleware,
+                                     accept_federated_only=self.federated_only,  # TODO: 466
+                                     certificate_filepath=certificate_filepath)
+                    self.log.debug("Verified node: {}".format(node.checksum_public_address))
+
+                else:
+                    node.validate_metadata(accept_federated_only=self.federated_only)  # TODO: 466
+
+            except node.SuspiciousActivity:
+                # TODO: Account for possibility that stamp, rather than interface, was bad.
+                message = "Suspicious Activity: Discovered node with bad signature: {}.  " \
+                          "Propagated by: {}".format(current_teacher.checksum_public_address, rest_url)
+                self.log.warn(message)
+            self.log.info("Previously unknown node: {}".format(node.checksum_public_address))
+            self.remember_node(node)
+            new_nodes.append(node)
+
+        self._adjust_learning(new_nodes)
+
+        learning_round_log_message = "Learning round {}.  Teacher: {} knew about {} nodes, {} were new."
+        current_teacher.last_seen = maya.now()
+        self.cycle_teacher_node()
+        self.log.info(learning_round_log_message.format(self._learning_round,
+                                                        current_teacher.checksum_public_address,
+                                                        len(node_list),
+                                                        len(new_nodes)), )
+        if new_nodes and self.known_certificates_dir:
+            for node in new_nodes:
+                node.save_certificate_to_disk(self.known_certificates_dir, force=True)
+
+        return new_nodes
 
 
 class VerifiableNode:
